kuber_hex v0.0.2 Kuber.Hex.Gateways.AuthorizeNet

Functions

authenticate(opts)

function to authorize the merchant using merchant name and transactionKey

authorize(amount, payment, opts)

Use this method to authorize a card payment. To actually charge funds, a follow up with capture method needs to be done.

capture(id, opts)

capture(amount, id, opts)

Use this method to capture funds for transactions which have been authorized, requires transId of the authorize function response to be passed as id along with the amount to be captured.

join_string(list, symbol)

purchase(amount, payment, opts)

Function to charge a user credit card for the specified amount, according to the payment method provided(e.g. credit card, apple pay etc.)

refund(amount, id, opts)

Use this method to refund a customer for a transaction that was already settled

store(card, opts)

unstore(customer_id, card_id, opts)

validate_config(config)

Validates the config dynamically depending on what is the value of required_config

void(id, opts)

Use this method to cancel either an original transaction that is not settled or an entire order composed of more than one transaction. It can be submitted against any other transaction type. Requires the transId of a request passed as id.