kuber_hex v0.0.2 Kuber.Hex.Gateways.Cams

An API client for the CAMS gateway. For referance you can test gateway operations [CAMS API SANDBOX] (https://secure.centralams.com).

Test it using test crediantials username:”testintegrationc”, password:”password9”

The following features of CAMS are implemented:

Purchase In purchase/3

Authorize In authorize/3

Capture In capture/3

Refund In refund/3

Void In void/2

Verify In verify/2

Functions

authorize(money, payment, options)

authorize(number(), Kuber.Hex.CreditCard, Keyword) :: Kuber.Hex.Response

authorize method responsible for authorize the credit card with specified amount. Authorize method authorize the transaction but does not trasnfer the funds. After authorize we need to call capture method for complete the transaction(fund transfer). After successful authorisation it returns response containing authcode & transactionid. auth_code and transactionid is required for capturing specified amount for particular transaction. After successful transaction it returns response with authcode & transactionid

capture(id, opts)

capture(money, authorization, options)

capture(number(), String.t(), Keyword) :: Kuber.Hex.Response

An authorized transaction need to be captured to compelete the transaction.

purchase(money, payment, options)

purchase(number(), Kuber.Hex.CreditCard, Keyword) :: Kuber.Hex.Response

Purchase method responsible for sale operation. After successfull transaction it returns response containing authcode & transactionid

refund(money, authorization, options)

refund(number(), String.t(), Keyword) :: Kuber.Hex.Response

Successful trasaction can be refunded after settlement or any time. It require transactionid for refund the specified amount back to creditcard holder.

store(card, opts)

unstore(customer_id, card_id, opts)

validate_config(config)

Validates the config dynamically depending on what is the value of required_config

verify(credit_card, options)

verify(Kuber.Hex.CreditCard, Keyword) :: Kuber.Hex.Response

This mehthod is for only verification purpose for credit card. Here we need to provide CreditCard info.

void(authorization, options)

void(String.t(), Keyword) :: Kuber.Hex.Response

Transaction can be cancled only before settlement. It require transactionid for cancel the transaction.