kuber_hex v0.0.2 Kuber.Hex.Gateways.Monei

An API client for the MONEI gateway.

For reference see MONEI’s API (v1) documentation.

The following features of MONEI are implemented:

PA Pre-Authorize
In authorize/3.

CP Capture
In capture/3.

RF Refund
In refund/3 and also void/2.

RV Reversal
In void/2.

DB Debit
In purchase/3.

Tokenization and Registrations
In store/2 and unstore/2.

The opts argument

This is a Keyword list of optional arguments for transactions with the MONEI gateway. The following keys are supported:

Key	Remark	Status

billing_address		Not implemented

shipping_address		Not implemented

customer		Not implemented

shipping_customer		Not implemented

merchant		Not implemented

cart		Not implemented

invoice		Not implemented

customParameters		Not implemented

currency		Not implemented

MONEI quirks

MONEI does not process money in cents, and the amount is rounded to 2 decimal places.

Caveats

Although MONEI supports payments from various cards, banks and virtual accounts (like some wallets), this library only accepts payments by (supported) cards.

TODO

Backoffice operations

Credit

Rebill

Recurring payments

Reporting

Functions

authorize(amount, card, opts)

authorize(number(), Kuber.Hex.CreditCard, keyword()) :: Kuber.Hex.Response

Performs a (pre) Authorize operation.

The authorization validates the card details with the banking network, places a hold on the transaction amount in the customer’s issuing bank also triggers risk management. Funds are not transferred.

MONEI returns an ID string which can be used to:

capture/3 an amount.

void/2 a pre-authorization.

Note

A stand-alone pre-authorization expires in 72hrs.

capture(amount, paymentId, opts)

capture(number(), String.t(), keyword()) :: Kuber.Hex.Response

Captures a pre-authorized amount.

amount is transferred to the merchant account by MONEI when it is smaller or equal to the amount used in the pre-authorization referenced by paymentId.

Note

MONEI allows partial captures and unlike many other gateways, does not release the remaining amount back to the payment source. Thus, the same pre-authorisation ID can be used to perform multiple captures, till:

all the pre-authorized amount is captured or,

the remaining amount is explicitly “reversed” via void/2. [citation-needed]

commit(method, endpoint, params, opts)

commit(atom(), String.t(), keyword(), keyword()) ::
  {:ok, HTTPoison.Response} |
  {:error, HTTPoison.Error}

Makes the request to MONEI’s network.

purchase(amount, card, opts)

purchase(number(), Kuber.Hex.CreditCard, keyword()) :: Kuber.Hex.Response

Credits the merchant account with amount by debiting the account of the customer.

MONEI attempts to debit the customer to accept a payment of amount with the given card.

refund(amount, paymentId, opts)

refund(number(), String.t(), keyword()) :: Kuber.Hex.Response

Credits the account of the customer with a reference to a prior transfer.

MONEI can process a full or partial refund worth amount, referencing a previous purchase/3 or capture/3ed.

The end customer will always see two bookings/records on his statement. Refer MONEI’s Backoffice Operations guide.

respond(arg)

Parses MONEI’s response and returns a Response struct in a :ok, :error tuple.

store(card, opts)

store(Kuber.Hex.CreditCard, keyword()) :: Kuber.Hex.Response

Stores the payment-source data for later use.

MONEI can store the payment-source details, for example card or bank details which can be used to effectively process One-Click and Recurring payments, and return a registration token for reference.

It is recommended to associate these details with a “Customer” by passing customer details in the opts.

Note

One-Click and Recurring payments are currently not implemented.

Payment details can be saved during a purchase/3 or capture/3.

unstore(arg, opts)

unstore(String.t(), keyword()) :: Kuber.Hex.Response

WIP

MONEI unstore does not seem to work. MONEI always returns a 403

Deletes previously stored payment-source data.

validate_config(config)

void(paymentId, opts)

void(String.t(), keyword()) :: Kuber.Hex.Response

Voids the referenced payment.

This method attempts a reversal of the paymentId referencing either a previous purchase/3 or authorize/3.

Voiding a previous authorization

MONEI will reverse the authorization by sending a “reversal request” to be sent the payment source (card issuer) to clear the funds held against the authorization. If some of the authorized amount was captured, only the remaining amount is cleared. [citation-needed]

Voiding a previous purchase

MONEI will reverse the payment, by sending all the amount back to the customer.

As a consequence, the customer will never see any booking on his statement. Refer MONEI’s Backoffice Operations guide.