kuber_hex v0.0.2 Kuber.Hex.Gateways.Stripe

Functions for working with Stripe payment gateway. Through this API you can:

Authorize payment source and use it later for payment.

Purchase with payment source.

Capture a payment for already authorized payment source.

Void the payment for payment source.

Refund amount to payment source.

Store payment source for making purchases later.

Unstore payment source.

Stripe API reference: https://stripe.com/docs/api

Functions

authorize(amount, payment, opts)

authorize(Float, Map, List) :: Map

Authorize payment source.

Authorize the payment source for a customer or card using amount and opts. opts must include the default currency.

Examples

payment = %{

expiration: {2018, 12}, number: "4242424242424242", cvc:  "123", name: "John Doe",
street1: "123 Main", street2: "Suite 100", city: "New York", region: "NY", country: "US",
postal_code: "11111"

}

opts = [currency: “usd”] amount = 5

Kuber.Hex.Gateways.Stripe.authorize(amount, payment, opts)

capture(id, opts)

capture(id, amount, opts \\ [])

capture(String.t(), Float, List) :: Map

Captures a payment.

Captures a payment with already authorized payment source. Once the charge is captured, it cannot be captured again. Amount less than or equal to authorized amount can be captured but not more than that. If less amount is captured than the authorized amount, then remaining amount will be refunded back to the authorized paymet source.

Examples

id = “ch_1BYvGkBImdnrXiZwet3aKkQE” amount = 5 opts = []

Kuber.Hex.Gateways.Stripe.capture(id, amount, opts)

purchase(amount, payment, opts)

purchase(Float, Map, List) :: Map

Purchase with payment source.

Purchase with the payment source using amount and opts. opts must include the default currency.

Examples

payemnt = %{

expiration: {2018, 12}, number: "4242424242424242", cvc:  "123", name: "John Doe",
street1: "123 Main", street2: "Suite 100", city: "New York", region: "NY", country: "US",
postal_code: "11111"

}

opts = [currency: “usd”] amount = 5

Kuber.Hex.Gateways.Stripe.purchase(amount, payment, opts)

refund(amount, id, opts)

refund(Float, String.t(), List) :: Map

Refunds the amount.

Returns the captured amount to the authorized payment source. Less than or equal to the captured amount can be refunded. If the refunded amount is less than the captured amount, then remaining amount can be refunded again.

Examples

amount = 5 id = “ch_1BYvGkBImdnrXiZwet3aKkQE” opts = []

Kuber.Hex.Gateways.Stripe.refund(amount, id, opts)

store(payment, opts)

store(Map, List) :: Map

Stores the payment source.

Store the payment source, so that it can be used for capturing the amount at later stages.

Examples

payment = %{

expiration: {2018, 12}, number: "4242424242424242", cvc:  "123", name: "John Doe",
street1: "123 Main", street2: "Suite 100", city: "New York", region: "NY", country: "US",
postal_code: "11111"

}

opts = []

Kuber.Hex.Gateways.Stripe.store(payment, opts)

unstore(id, opts \\ [])

Unstore the stored payment source.

Unstore the already stored payment source, so that it cannot be used again for capturing payments.

Examples

id = “cus_BwpLX2x4ecEUgD”

Kuber.Hex.Gateways.Stripe.unstore(id)

unstore(customer_id, card_id, opts)

validate_config(config)

Validates the config dynamically depending on what is the value of required_config

void(id, opts)

void(String.t(), List) :: Map

Voids the payment.

Returns the captured amount to the authorized payment source.

Examples

id = “ch_1BYvGkBImdnrXiZwet3aKkQE” opts = []

Kuber.Hex.Gateways.Stripe.void(id, opts)