kuber_hex v0.0.2 Kuber.Hex.Response

Module which defines the struct for response struct.

Response struct is a standard response from public API to the application.

TODO: Add the response struct detail about the attributes in the struct

It mostly has such as:-

* `success`: boolean indicating the status of the transaction
* `authorization`: token which is used to issue requests without the card info
* `code`: status code for the response
* `reason`: reason for the error if it happens
* `avs_result`: TODO: add this
* `cvc_result`: result for cvc verification 
* `raw`: TODO: add this

Functions

error(opts \\ [])

success(opts \\ [])