kuber_hex v0.0.2 Kuber.Hex.Gateways.Cams

An API client for the CAMS gateway. For referance you can test gateway operations [CAMS API SANDBOX] (https://secure.centralams.com).

Test it using test crediantials username:testintegrationc, password:password9

The following features of CAMS are implemented:

Purchase In purchase/3

Authorize In authorize/3

Capture In capture/3

Refund In refund/3

Void In void/2

Verify In verify/2

Functions

authorize(money, payment, options)

authorize(number(), Kuber.Hex.CreditCard, Keyword) :: Kuber.Hex.Response

Use this method for authorising the credit card for particular transaction.

authorize/3 method only authorise the transaction,it does not transfer the funds. After authorised a transaction, we need to call capture/3 method to complete the transaction. After successful authorisation it returns response containing authcode and transactionid. We required authcode,transactionid and money for capturing transaction later on. It perform operation by taking money, payment (credit card details) & options as parameters. Here money is required field which contains amount to be deducted. Required fields in credit card are Credit Card Number & Expiration Date. Whereas options contains other information like billing address,order information etc.

capture(id, opts)

capture(money, authorization, options)

capture(number(), String.t(), Keyword) :: Kuber.Hex.Response

Use this method for capture the amount of the authorised transaction which is previously authorised by authorize/3 method.

It take money, authorization and options as parameters. Where money is a amount to be captured and authorization is a response returned by authorize/3 method. From response it takes authcode and transactionid for further processing. Both money and authorization are required fields, whereas options are as same as authorize/3 and purchase/3 methods.

purchase(money, payment, options)

purchase(number(), Kuber.Hex.CreditCard, Keyword) :: Kuber.Hex.Response

Use this method for performing purchase(sale) operation.

It perform operation by taking money, payment(credit card details) & options as parameters. Here money is required field which contains amount to be deducted. Required fields in credit card are Credit Card Number & Expiration Date. Whereas options contains other information like billing address,order information etc. After successful transaction it returns response containing authcode & transactionid.

refund(money, authorization, options)

refund(number(), String.t(), Keyword) :: Kuber.Hex.Response

Use this method for refund the amount for particular transaction.

Successful transaction can be refunded after settlement or any time. It requires transactionid for refund the specified amount back to credit card holder. Only purchased(sale) transactions can be refund based on thier transactionid. It take money, authorization and options as parameters. Where money is a amount to be refund and authorization is a response returned by purchase/3 method. From response it takes authcode and transactionid for further processing. Both money and authorization are required fields, whereas options are as same as authorize/3, purchase/3 and capture/3 methods.

store(card, opts)

unstore(customer_id, card_id, opts)

validate_config(config)

Validates the config dynamically depending on what is the value of required_config

verify(credit_card, options)

verify(Kuber.Hex.CreditCard, Keyword) :: Kuber.Hex.Response

This mehthod is for only verification purpose for credit card. Here we need to provide CreditCard info.

void(authorization, options)

void(String.t(), Keyword) :: Kuber.Hex.Response

Transaction can be cancled only before settlement. It require transactionid for cancel the transaction.